moved unit test functions into utils
